move accordion code to `source-react-components`
