+++ packages/@guardian/source-react-components/src/accordion/theme.ts
@@ -1,0 +1,8 @@
+import { border, text } from '@guardian/source-foundations';
+
+export const accordionThemeDefault = {
+	accordion: {
+		textPrimary: text.primary,
+		borderPrimary: border.primary,
+	},
+};
