--- packages/@guardian/source-react-components/src/stories/Accordion.stories.tsx
@@ -1,8 +1,0 @@
-import defaultStoryConfig from '../../../src-accordion/Accordion.stories';
-
-export default {
-	...defaultStoryConfig,
-	title: 'Source v4/source-react-components/Accordion',
-};
-
-export * from '../../../src-accordion/Accordion.stories';
