--- packages/@guardian/src-accordion/.yarnrc
@@ -1,2 +1,0 @@
-version-tag-prefix src-accordion-v
-version-git-message "@guardian/src-accordion v%s"
