--- packages/@guardian/src-accordion/README.md
@@ -1,17 +1,0 @@
-# `@guardian/src-accordion`
-
-## Install
-
-```sh
-$ yarn add @guardian/src-accordion @guardian/src-foundations
-```
-
-## Use
-
-### API
-
-See [storybook](https://guardian.github.io/source/?path=/docs/source-src-accordion-accordion--demo).
-
-### How to use
-
-For context and visual guides relating to accordion usage see the [Source Design System website](https://www.theguardian.design/2a1e5182b/p/38c5aa-accordion).
