--- packages/@guardian/src-accordion/index.tsx
@@ -1,5 +1,0 @@
-export { Accordion } from './Accordion';
-export type { AccordionProps } from './Accordion';
-export { AccordionRow } from './AccordionRow';
-export type { AccordionRowProps } from './AccordionRow';
-export { accordionDefault } from '@guardian/src-foundations/themes';
