--- packages/@guardian/src-accordion/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "@guardian/src-accordion",
-  "version": "3.12.1",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/guardian/source.git"
-  },
-  "license": "Apache-2.0",
-  "main": "dist/cjs/index.js",
-  "module": "dist/esm/index.js",
-  "types": "dist/types/index.d.ts",
-  "files": [
-    "dist/esm",
-    "dist/cjs",
-    "dist/types"
-  ],
-  "scripts": {
-    "build": "npm-run-all --parallel build:*",
-    "build:cjs": "tsc --composite false --declaration false --emitDeclarationOnly false --outDir dist/cjs --module commonjs",
-    "build:esm": "tsc --composite false --declaration false --emitDeclarationOnly false --outDir dist/esm --module ES6",
-    "build:types": "tsc -b",
-    "clean": "rm -rf dist",
-    "publish:public": "yarn publish --access public",
-    "verbump:major": "yarn version --major --no-git-tag-version",
-    "verbump:minor": "yarn version --minor --no-git-tag-version",
-    "verbump:patch": "yarn version --patch --no-git-tag-version",
-    "verbump:premajor": "yarn version --premajor --preid rc --no-git-tag-version",
-    "verbump:preminor": "yarn version --preminor --preid rc --no-git-tag-version",
-    "verbump:prerelease": "yarn version --prerelease --preid rc --no-git-tag-version"
-  },
-  "dependencies": {
-    "@guardian/libs": "^3.2.1",
-    "@guardian/src-helpers": "^3.12.1"
-  },
-  "devDependencies": {
-    "npm-run-all": "^4.1.5",
-    "typescript": "^4.3.0"
-  },
-  "peerDependencies": {
-    "@emotion/react": "^11.1.2",
-    "@guardian/src-foundations": "^3.10.3-rc.1",
-    "@guardian/src-icons": "^3.10.3-rc.1",
-    "react": "^17.0.1"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
