--- packages/@guardian/src-accordion/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "extends": "../../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist/types",
-    "declaration": true,
-    "emitDeclarationOnly": true
-  },
-  "include": [".", "../src-foundations/types/themes.d.ts"],
-  "exclude": ["dist", "*.stories.tsx"],
-  "references": [
-    {
-      "path": "../src-foundations"
-    },
-    {
-      "path": "../src-icons"
-    }
-  ]
-}
