--- packages/@guardian/src-foundations/src/themes/accordion.ts
@@ -1,8 +1,0 @@
-import { text, border } from '@guardian/src-foundations/palette';
-
-export const accordionDefault = {
-	accordion: {
-		textPrimary: text.primary,
-		borderPrimary: border.primary,
-	},
-};
