+++ packages/@guardian/src-foundations/src/themes/index.ts
@@ -1,4 +1,3 @@
-export * from './accordion';
 export * from './button';
 export * from './checkbox';
 export * from './choice-card';
